Update the pipline - Enable new blocking pass and remove the old one - Remove the PropagationLayout Pass
