+++ lib/Dialect/XeTile/Transforms/WgToSg.cpp
@@ -40,7 +40,6 @@
 #include <cassert>
 #include <optional>
 
-#include "imex/Dialect/XeTile/Transforms/Blocking.h"
 #include "imex/Dialect/XeTile/Transforms/Passes.h"
 #include "imex/Utils/DebugUtils.h"
 #include "imex/Utils/XeArch.h"
